kern: compute MPU contents at build time.
This reduces the cost of the `apply_memory_protection` phase of context
switches by computing the descriptors in a const fn and sneaking them
into the RegionDesc struct via a new field.

This causes MPU load on ARMv7-M (Cortex-M4) to finish in 45% the time it
previously took, which in turn knocks about 266 cycles off IPC-related
syscalls.