arm_m: reduce scheduler overhead for ISRs
Currently, the default ISR implementation for ARM-M does not do an
automatic context save for the interrupted task. This is under the
assumption that many ISRs will happen during higher-priority tasks or
with the handling task not actively waiting, and so may not immediately
cause a context switch. When an ISR wants to trigger a context switch, it
does so using the PendSV mechanism, which is explained in the module
comments.

The way this was implemented had two performance drawbacks.

First, the ISR would trigger a PendSV (which costs at least a hundred
cycles or so) if the handling task was waiting, even if it was low
priority and would not immediately be a candidate for execution.

Second, while the ISR knows which task it's trying to activate, it
didn't share this information with PendSV, since there was no way to do
so. This meant PendSV had to invoke the scheduler to choose a new task,
which is more expensive than a directed switch into a specific task.

This change creates a communication channel from ISRs to PendSV through
a global variable. An ISR can nominate a "proposed" next task, and
PendSV will consider it. If the proposed next task is higher priority
than the current task, it gets activated directly. If not, the old
behavior kicks in.

In addition, this changes DefaultHandler to check priorities before
poking PendSV, as a relatively inexpensive defense against unnecessary
PendSV exceptions.

In a very interrupt-heavy demo application (15k IRQs/s, 175k context
switches/s), this reduces overall CPU utilization by about 4%.