kern: move task table guard into BSS
This variable was previously interpreted in a way that required it to be
initialized to true, then overridden to false. This meant it wound up in
a .data section instead of .bss ... the only such variable in the
kernel. Because of alignment requirements on the bss section this caused
us to lose 7 bytes to padding.

Flipping the polarity of the variable causes it to move to BSS and pack
nicely with all the other variables.