kern: move task table (both its decl, and its location)
The task table has historically been in BSS, so it's zeroed by the
kernel startup routine before being overwritten. Because it's a
MaybeUninit, we don't actually have to do this; there's no risk of us
reading the memory before we write it, and we diligently write it during
kernel startup.

The task table has _also_ historically been declared by generated code.
This is a bit obscure and makes it difficult to comment.

This change moves the task table's declaration into
`kern/src/startup.rs` next to everything else related to startup, and
moves its placement by the linker script into the .uninit section.

Since the task table is the _vast majority_ of kernel RAM (95% in the
simplest applications, proportionally more as tasks are added), this
removes the vast majority of time spent zeroing memory during
application startup.