arm_m: try improve armv8m context switch cache performance further
Instead of saving a single u8 in each RegionDescExt which ends up
growing the struct by 4 bytes, store two u32s in a TaskDescExt struct
that expands the TaskDesc struct with architecture specific data.

The main downside here is that TaskDesc is 64 bits larger on armv8m.
If necessary, it's possible to move the TaskDescExt onto a separate
array with some extra changes.